Go through 2018/ferguson and 2020/ferguson[12]
Surprise, surprise there are more updates on my entries! :-) But to be
fair to me there were some important fixes and I had not really gone
through the README files of these entries (I still didn't go through all
of them thoroughly and probably won't since I'm far too concise :-) :-)).
In fact some of these problems broke some of the features of some of
the entries.

I do want to note a few fixes that are important to consider. In
2020/ferguson[12] there is an EXTRA_PROG variable. I'm not sure if this
is the right name but it's not an alt target: they're extra tools that
are there. For instance in 2020/ferguson2 the recode program was not
compiled which is rather a problem. The same could be said for
2020/ferguson1/termcaps: that it was not compiled by make broke some
scripts.

There were some typo fixes here and there as well as (I hope) a bit
clearer wording.

Some out of date things were changed.

As for 2018/ferguson an important point in the Makefile (besides adding
the ALT_TARGET that should have been there) is that the Makefile makes a
symlink to prog as weasel. This is because the judges originally did
this and to not have it would take fixing a lot in the documentation.
Now Landon wrote to me the other day:

    We suspect you might never be finished with your write-up, @xexyl,
but that is OK as the new system is designed to give author's the
ability to improve their write-ups via GitHub Pull requests ...

    ... as long as they don't happen too often 😁

...which if I were to fix this there would be yet another update which
isn't very desirable, right? :-) Nevertheless if necessary I will go
through the files and see if I can find the correct references to weasel
to change to prog (unfortunately it's not a simple matter of a global
substitution).