More typo fixes in tmp/example.dot_winner.json
I believe that the chocolate_cake.jpg (and I'm super honoured and feel
privileged that it was put into the proof of concept file even if proof
of concept has the abbreviation of another less than rewarding thing :-)
:-) ) should not be a required argument but rather a required data file
(and it certainly should be required !! :-) ).

Similarly raw_data.file should not be an optional argument but
optional data file.