Address 2011/richards in bugs.md wrt Apple M1
I've kind of wondered for a while if it's due to Apple protections
against memory protection and it seems like that this might be the case
either for the Apple silicon chip or others too (I can only test the M1
chip).

I added resources on Apple's website as well as adding a bit of things I
discovered but I think at this time this cannot be resolved easily or at
all due to Apple labouring under the delusion (as the expression goes
even if it's not strictly accurate) that they 'know better than everyone
else'. I will proceed to other years now, starting with 2013/birken, in
the next commits (unless I find a typo or something in an earlier year
prior to continuing 2013 :-) ).