Fix paths in contact.md and contact.html
Changed ../judges.html to /judges.html and faq.html to /faq.html. These
are of course pedantic fixes: for of course at the root directory
../judges.html will evaluate to /judges.html and at the root directory
faq.html will be the same as /faq.html so these updates don't fix a
problem but rather make it more correct. The reason for the / is in case
for some reason any of the files are moved to subdirectory (or if the
files are in different directories, I guess), unlikely as that is. This
was done in other files quite some time back and probably should be done
throughout the repo but that can be dealt with another time.