Fix 'web site' to 'website'
This is a huge update with 470 files updated. This includes the scripts
and config files that generate the html files that have (now) 'Fix the
website' (was 'Fix the web site') but mostly the html files that were
generated from scripts and markdown files. This is the only change in
the commit so although there are a lot of files to look at in the diff
it should be easy to compare.

The rationale, for those who might be curious: there was an
inconsistency where in some files it was 'web site' (most files) and
others 'website' (those by me). Since this was inconsistent I brought it
up on GitHub and it was suggested that it should all be 'website'.