Change references of 'web site' to 'website'
As agreed on GitHub to be consistent this should be done. This commit
changes 470 files. There are some files that have not had this change.
Some it might be okay to change but there is one that should certainly
not be changed. The files not changed are thus:

    2004/hibachi/src/localhost/index.htm

This could be changed but it is part of the entry is not strictly
necessary.

    2011/dlowe/dlowe-aux-data/english-1/Wireless_Days_2011_CFP.txt

This should not be changed because it is part of the training data that
was used in the entry.

Any files under tmp/ that also refer to 'web site' have not been changed
as these files are going away.